requirements.txt
streamlit
pymupdf
python-docx
pandas
chromadb
langchain
sentence-transformers
